rename leading section to "Preface"
A foreword (not "foreward" or "forward", btw.) is written by someone
other than the author(s) of the book, usually by someone with
name recognition, to give the author(s) and the book credibility.

A frontmatter section by the author(s) that describes how the book came
to be or what the motivation behind making it was is called "preface".
This is what the content of `ch01-00-*.md` currently is.

Cf. https://en.wikipedia.org/wiki/Book_design#Front_matter